+++ richat/src/pubsub/server.rs
@@ -24,8 +24,8 @@ use {
         rt::tokio::{TokioExecutor, TokioIo},
         server::conn::auto::Builder as ServerBuilder,
     },
-    jsonrpsee_types::{ResponsePayload, TwoPointZero},
-    richat_shared::{jsonrpc::helpers::get_x_subscription_id, shutdown::Shutdown},
+    jsonrpsee_types::{Extensions, ResponsePayload, TwoPointZero},
+    richat_shared::jsonrpc::helpers::get_x_subscription_id,
     solana_nohash_hasher::IntMap,
     solana_rpc_client_api::response::RpcVersionInfo,
     std::{future::Future, net::TcpListener as StdTcpListener, sync::Arc},
@@ -34,6 +34,7 @@ use {
         sync::{broadcast, mpsc, oneshot},
     },
     tokio_rustls::TlsAcceptor,
+    tokio_util::sync::CancellationToken,
     tracing::{error, info, warn},
 };
 
@@ -44,7 +45,7 @@ impl PubSubServer {
     pub fn spawn(
         mut config: ConfigAppsPubsub,
         messages: Messages,
-        shutdown: Shutdown,
+        shutdown: CancellationToken,
     ) -> anyhow::Result<impl Future<Output = anyhow::Result<()>>> {
         let acceptor = config
             .tls_config
@@ -95,7 +96,6 @@ impl PubSubServer {
         // Spawn server
         let server_jh = tokio::spawn(async move {
             let mut client_id = 0;
-            tokio::pin!(shutdown);
             loop {
                 // accept connection
                 let stream = tokio::select! {
@@ -112,7 +112,7 @@ impl PubSubServer {
                             break;
                         }
                     },
-                    () = &mut shutdown => break,
+                    () = shutdown.cancelled() => break,
                 };
 
                 // Create service
@@ -222,7 +222,7 @@ impl PubSubServer {
         enable_token_init_subscription: bool,
         clients_tx: mpsc::Sender<ClientRequest>,
         mut notifications: broadcast::Receiver<RpcNotification>,
-        shutdown: Shutdown,
+        shutdown: CancellationToken,
     ) -> anyhow::Result<()> {
         let mut ws = ws_fut.await?;
         ws.set_max_message_size(recv_max_message_size);
@@ -236,7 +236,6 @@ impl PubSubServer {
             let mut send_frame = None;
             let mut last_frame = false;
             let mut send_fn = |_| async { Ok::<(), String>(()) };
-            tokio::pin!(shutdown);
             loop {
                 if let Some(frame) = send_frame.take() {
                     let (tx, rx) = oneshot::channel();
@@ -252,7 +251,7 @@ impl PubSubServer {
                 // read msg
                 let frame = tokio::select! {
                     frame = ws_rx.read_frame(&mut send_fn) => frame?,
-                    () = &mut shutdown => break,
+                    () = shutdown.cancelled() => break,
                 };
                 let payload = match frame.opcode {
                     OpCode::Close => {
@@ -363,7 +362,8 @@ impl PubSubServer {
                             let vec = serde_json::to_vec(&jsonrpsee_types::Response {
                                 jsonrpc: Some(TwoPointZero),
                                 payload: ResponsePayload::success(result),
-                                id: message.id
+                                id: message.id,
+                                extensions: Extensions::default(), // doesn't matter, as it is not used in serialize
                             }).expect("json serialization never fail");
                             let frame = Frame::text(Payload::Owned(vec));
                             ws_tx.write_frame(frame).await?;